+++ R/edgebundle-package.R
@@ -5,10 +5,11 @@
 #' An R package that implements several edge bundling/flow and metro map algorithms.
 #' It includes:
 #'
-#' - Force directed edge bundling `edge_bundle_force()`
+#' - Force directed edge bundling `edge_bundle_force()` (with a divided/directed variant)
 #' - Stub bundling  `edge_bundle_stub()`
-#' - Hammer bundling  `edge_bundle_hammer()`
+#' - Hammer bundling (KDE) `edge_bundle_hammer()`
 #' - Edge-path bundling  `edge_bundle_path()`
+#' - MINGLE bundling  `edge_bundle_mingle()`
 #' - TNSS flow map `tnss_*()`
 #' - Multicriteria Metro map layout `metro_multicriteria()`
 #'